+++ SQL_Server/DTOs/AdminPhoneDTO.cs
@@ -2,7 +2,7 @@ namespace SQL_Server.DTOs
 {
     public class AdminPhoneDTO
     {
-        public required int Admin_id { get; set; }
-        public required int Phone { get; set; }
+        public required long Admin_id { get; set; }
+        public required long Phone { get; set; }
     }
 }
