+++ SQL_Server/DTOs/BusinessAssociateDTO.cs
@@ -2,7 +2,7 @@ namespace SQL_Server.DTOs
 {
     public class BusinessAssociateDTO
     {
-        public required int Legal_Id { get; set; } // PK
+        public required long Legal_Id { get; set; } // PK
         public required string Email { get; set; }
         public required string State { get; set; } // "Aceptado", "En espera", "Rechazado"
         public required string BusinessName { get; set; }
@@ -10,9 +10,9 @@ namespace SQL_Server.DTOs
         public required string Province { get; set; }
         public required string Canton { get; set; }
         public required string District { get; set; }
-        public required int SINPE { get; set; }
+        public required long SINPE { get; set; }
         public string? RejectReason { get; set; }
         public required string BusinessManager_Email { get; set; } // FK
-        public required int BusinessType_Identification { get; set; } // FK
+        public required long BusinessType_Identification { get; set; } // FK
     }
 }
