+++ SQL_Server/DTOs/BusinessManagerPhoneDTO_Update.cs
@@ -2,6 +2,6 @@ namespace SQL_Server.DTOs
 {
     public class BusinessManagerPhoneDTO_Update
     {
-        public required int Phone { get; set; }
+        public required long Phone { get; set; }
     }
 }
