+++ SQL_Server/DTOs/Cart_ProductDTO.cs
@@ -2,8 +2,8 @@ namespace SQL_Server.DTOs
 {
     public class Cart_ProductDTO
     {
-        public required int Cart_Code { get; set; } // PK, FK
-        public required int Product_Code { get; set; } // PK, FK
-        public required int Amount { get; set; }
+        public required long Cart_Code { get; set; } // PK, FK
+        public required long Product_Code { get; set; } // PK, FK
+        public required long Amount { get; set; }
     }
 }
