+++ SQL_Server/DTOs/Cart_ProductDTO_Create.cs
@@ -2,7 +2,7 @@ namespace SQL_Server.DTOs
 {
     public class Cart_ProductDTO_Create
     {
-        public required int Cart_Code { get; set; } // FK
-        public required int Product_Code { get; set; } // FK
+        public required long Cart_Code { get; set; } // FK
+        public required long Product_Code { get; set; } // FK
     }
 }
