+++ SQL_Server/DTOs/Cart_ProductDTO_Update.cs
@@ -2,6 +2,6 @@ namespace SQL_Server.DTOs
 {
     public class Cart_ProductDTO_Update
     {
-        public required int Amount { get; set; }
+        public required long Amount { get; set; }
     }
 }
